Refatorar inicialização do aplicativo: remover arquivo Inicializador.dart, atualizar main.dart para incluir tema e dependências, e ajustar importações nos testes
