--- REPRESENTANTE/pacsoft_representante/lib/Inicializador.dart
@@ -1,16 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:pacsoft_representante/HOME/INICIAL/PG_INICIAL.dart';
-import 'package:pacsoft_representante/HOME/LOGIN/PG_LOGIN.dart';
-
-
-
-class Pacsoft extends StatelessWidget {
-  const Pacsoft({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      home:  Login(),
-    );
-  }
-}
