+++ REPRESENTANTE/pacsoft_representante/lib/main.dart
@@ -1,7 +1,23 @@
 import 'package:flutter/material.dart';
-import 'package:pacsoft_representante/Inicializador.dart';
+import 'package:pacsoft_representante/HOME/LOGIN/PG_LOGIN.dart';
 
 void main() {
   WidgetsFlutterBinding.ensureInitialized();
   runApp(Pacsoft());
+}
+
+class Pacsoft extends StatelessWidget {
+  const Pacsoft({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      title: 'Pacsoft',
+      theme: ThemeData(
+        primarySwatch: Colors.green,
+        fontFamily: 'Lato',
+      ),
+      home:  Login(),
+    );
+  }
 }
