+++ REPRESENTANTE/pacsoft_representante/test/widget_test.dart
@@ -7,7 +7,7 @@
 
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:pacsoft_representante/Inicializador.dart';
+import 'package:pacsoft_representante/main.dart';
 
 
 void main() {
